Refactor (#15): extract buildStarElement/buildLinkElement helpers and adopt in mapState (no behavior change)
